+++ backend/models/transactionLogModel.js
@@ -17,7 +17,7 @@ TransactionLogModel.init(
       allowNull: false,
     },
     amount: {
-      type: DataTypes.FLOAT,
+      type: DataTypes.DECIMAL(12, 2),
       allowNull: false,
     },
     transaction_id: {
